[seller] feat(123) lint 에러 수정 및 FSD 규칙 준수 개선
