+++ apps/seller/src/features/products/create/create-form-delivery/create-form-delivery-area.ui.tsx
@@ -4,7 +4,7 @@ import { Dropdown, Input, Label } from '@dessert/ui'
 import { Controller } from 'react-hook-form'
 
 import { DeliveryCompany, DeliveryTerms } from '@/entity/products'
-import { InfoTooltip } from '@/features/products/create/ui/info-tooltip.ui'
+import { InfoTooltip } from '../create-form/info-tooltip.ui'
 import { cn } from '@/shared/libs/utils'
 
 import { useProductDeliveryForm } from './use-product-delivery-form.hook'
