+++ apps/seller/src/features/products/create/create-form-detail/index.ts
@@ -1,0 +1,1 @@
+export { ProductDetailArea } from './create-form-detail-area.ui'
