+++ apps/seller/src/features/products/create/create-form-disclosure/index.ts
@@ -1,0 +1,3 @@
+export { ProductDisclosureArea } from './create-form-disclosure-area.ui'
+export { useProductDisclosureForm } from './use-product-disclosure-form.hook'
+export { disclosureSchema } from './create-disclosure.schema'
