+++ apps/seller/src/features/products/create/create-form-info/index.ts
@@ -1,0 +1,3 @@
+export { ProductInfoArea } from './create-form-info-area.ui'
+export { useProductInfoForm } from './use-product-info-form.hook'
+export { productSchema } from './create-info.schema'
