+++ apps/seller/src/features/products/create/create-form-options/index.ts
@@ -1,0 +1,3 @@
+export { ProductOptionsArea } from './create-form-options-area.ui'
+export { useProductOptionForm } from './use-product-options.form.hook'
+export { productOptionSchema } from './create-options.schema'
