+++ apps/seller/src/features/products/create/create-header/index.ts
@@ -1,0 +1,2 @@
+export { ProductHeader } from './create-header.ui'
+export { ProductHeaderTags } from './create-header-tags.ui'
