+++ apps/seller/src/features/products/create/index.ts
@@ -1,20 +1,7 @@
-// Form Sections (UI)
-export { ProductDeliveryArea } from './create-form-delivery/create-form-delivery-area.ui'
-export { ProductDetailArea } from './create-form-detail/create-form-detail-area.ui'
-export { ProductDisclosureArea } from './create-form-disclosure/create-form-disclosure-area.ui'
-export { ProductInfoArea } from './create-form-info/create-form-info-area.ui'
-export { ProductOptionsArea } from './create-form-options/create-form-options-area.ui'
-
-// Header (UI)
-export { ProductHeader } from './create-header/create-header.ui'
-export { ProductHeaderTags } from './create-header/create-header-tags.ui'
-
-// Form Context & Hooks (Model/Lib)
-export { FormStepsProvider } from './create-form/create-form-provider.ui'
-export { useCreateFormSteps } from './create-form/use-create-form-steps.hook'
-
-// Schemas (Shared Logic)
-export { deliverySchema } from './create-form-delivery/create-delivery.schema'
-export { disclosureSchema } from './create-form-disclosure/create-disclosure.schema'
-export { productSchema } from './create-form-info/create-info.schema'
-export { productOptionSchema } from './create-form-options/create-options.schema'
+export * from './create-form'
+export * from './create-form-delivery'
+export * from './create-form-detail'
+export * from './create-form-disclosure'
+export * from './create-form-info'
+export * from './create-form-options'
+export * from './create-header'
